Criacao de novos controllers e do metodo DELETE no restaurante

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/RestauranteRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.algaworks.algafood.infrastructure.repository;
 
 import com.algaworks.algafood.domain.model.Restaurante;
 import com.algaworks.algafood.domain.repository.RestauranteRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,9 +34,15 @@ public class RestauranteRepositoryImpl implements RestauranteRepository {
 
     @Transactional
     @Override
-    public void remover(Restaurante restaurante) {
-       restaurante = buscar(restaurante.getId());
-       entityManager.remove(restaurante);
+    public void remover(Long id) {
+        Restaurante restaurante = buscar(id);
 
+        if(restaurante == null){
+            throw new EmptyResultDataAccessException(1);
+        }
+
+        entityManager.remove(restaurante);
     }
+
+
 }
